% Generated by roxygen2 (4.0.1): do not edit by hand
\name{labelCenters}
\alias{labelCenters}
\title{Label centers, i.e. assign risk categories to all centers}
\usage{
labelCenters(standardizedRisks, lambda = NULL, confidenceLevel = 0.5)
}
\arguments{
\item{standardizedRisks}{object of class 'standardizedRisk' as produced by the 'standardizeRisks' function}

\item{lambda}{vector of length two; the first element indicates the clinical tolerance
level before flagging as low mortality risk; the second element indicates
clinical tolerance level before flagging as high mortality risk; if NULL default values
of \code{c(low = 0.2, high = 0.2)} are used for direct standardization; for indirect
standardization a default value of \code{c(low = 0.05, high = 0.05)} is used}

\item{confidenceLevel}{statistical confidence level; default value 0.5}
}
\value{
object of class 'labeledCenters', which is a data frame with columns
  centerName, centerLabel, lowerCI and upperCI
}
\description{
Label centers, i.e. assign risk categories to all centers
}
\examples{
set.seed(130513)
simulatedData <- simulateData()
standardizedRisks <- with(simulatedData, standardizeRisks(patientCovariates = L, center = center, Y = Y))
labeledCenters <- labelCenters(standardizedRisks = standardizedRisks)
head(labeledCenters)
}

