% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{largeCaseMix}
\alias{largeCaseMix}
\title{Large Case-Mix Data}
\format{Data frame with the 5 variables:
\describe{
  \item{age}{age of the patient}
  \item{sex}{gender of the patient}
  \item{cons}{consciousness}
  \item{center}{center of the patient}
  \item{outcome}{binary outcome variable}
}}
\description{
Large Case-Mix Data
}
\examples{
data(largeCaseMix)
str(largeCaseMix)
}

