% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{largeCaseMix_missing}
\alias{largeCaseMix_missing}
\title{Large Case-Mix Data with Missing Consciousness Level for Certain Patients}
\format{Data frame with the 5 variables:
\describe{
  \item{age}{age of the patient}
  \item{sex}{gender of the patient}
  \item{cons}{consciousness}
  \item{center}{center of the patient}
  \item{outcome}{binary outcome variable}
}}
\description{
Large Case-Mix Data with Missing Consciousness Level for Certain Patients
}
\examples{
data(largeCaseMix_missing)
str(largeCaseMix_missing)
}

