% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCenterLabels}
\alias{plotCenterLabels}
\title{Plot the standardized risk with confidence limits per center}
\usage{
plotCenterLabels(standardizedRisks, labeledCenters, yValue = c("size",
  "name"), labelColors = c(L = "olivedrab1", A = "black", H = "red"),
  labelPch = c(L = 17, A = 20, H = 17), ...)
}
\arguments{
\item{standardizedRisks}{standardized risks as produced by the standardize function}

\item{labeledCenters}{labeled centers as produced by the labelCenters function}

\item{yValue}{value to report on the y axis; one of 'size' (default; center size) or
'name' (center name)}

\item{labelColors}{vector of colors (length 3) for the L, A and H label respectively;
default value is \code{c(L = "green", A = "black", H = "red")}}

\item{labelPch}{vector of plot characters to be used for the L, A and H label resp.;
default value is \code{c(L = 17, A = 20, H = 17)}}

\item{...}{further arguments passed to plot}
}
\value{
no return value; plot is written to the current device
}
\description{
Plot the standardized risk with confidence limits per center
}
\examples{
set.seed(130513)
simulatedData <- simulateData()
standardizedRisks <- with(simulatedData, standardizeRisks(patientCovariates = L, center = center, Y = Y))
labeledCenters <- labelCenters(standardizedRisks = standardizedRisks)
plotCenterLabels(standardizedRisks = standardizedRisks, labeledCenters = labeledCenters)
}

