% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotRisks}
\alias{plotRisks}
\title{Plot observed vs. expected risks}
\usage{
plotRisks(standardizedRisks, labeledCenters, labelColors = c(L = "olivedrab1",
  A = "black", H = "red"), labelPch = c(L = 17, A = 20, H = 17), ...)
}
\arguments{
\item{standardizedRisks}{object of class 'standardizedRisks' as produced by
by the standardizeRisks function}

\item{labeledCenters}{center labels as produced by the labelCenters function}

\item{labelColors}{vector of colors (length 3) for the L, A and H label respectively;
default value is \code{c(L = "green", A = "black", H = "red")}}

\item{labelPch}{vector of plot characters to be used for the L, A and H label resp.;
default value is \code{c(L = 17, A = 20, H = 17)}}

\item{...}{further arguments passed to plot and points}
}
\value{
no return value; a plot is written to the current device
}
\description{
Plot observed vs. expected risks
}
\examples{
set.seed(130513)
simulatedData <- simulateData()
standardizedRisks <- with(simulatedData, standardizeRisks(patientCovariates = L, center = center, Y = Y))
labeledCenters <- labelCenters(standardizedRisks = standardizedRisks)
plotRisks(standardizedRisks = standardizedRisks, labeledCenters = labeledCenters)
}
\seealso{
\link{standardizeRisks}
}

