% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulateData}
\alias{simulateData}
\title{Simulate example data for risk standardization}
\usage{
simulateData(n = 10000, m = 50, l = 5, betaL = FALSE)
}
\arguments{
\item{n}{total number of patients; default}

\item{m}{initial number of centers (min. 10); default value 50}

\item{l}{number of patient-specific covariates in outcome model; default}

\item{betaL}{if FALSE (default) L is distributed according to a N(0,1) distribution;
if TRUE L is distributed according to a Beta(1,6) distribution}
}
\value{
list with three components: L (model matrix), center (vector) and Y (vector; binary outcome variable)
}
\description{
Simulate example data for risk standardization
}
\examples{
set.seed(130513)
simulatedData <- simulateData()
str(simulatedData)
}

