% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardizeRisks}
\alias{standardizeRisks}
\title{Estimate standardized mortality risks}
\usage{
standardizeRisks(patientCovariates, center, Y, method = c("indirect",
  "direct"), Firth = TRUE, alpha = 0.05, missing = c("completeCase",
  "dummyCategory"), trace = FALSE)
}
\arguments{
\item{patientCovariates}{data frame of patient-specific covariates}

\item{center}{center code (n values for n patients)}

\item{Y}{binary outcome}

\item{method}{method of standardization; one of 'indirect' (default) or 'direct'}

\item{Firth}{logical apply Firth correction? default value TRUE}

\item{alpha}{statistical significance level; default value is 0.05}

\item{missing}{how to handle missing categorical data? one of 'completeCase' (default) or 'dummyCategory';
when 'dummyCategory' is chosen a separate category is added to model a missing value effect}

\item{trace}{logical print summary of fitted model? default value FALSE}
}
\value{
object of class 'standardizedRisks'; data frame with columns
centerName, centerSize, standardizedRisk, varStandardizedRisk, lowerCI, upperCI and observedRisk
}
\description{
Estimate standardized mortality risks
}
\examples{
set.seed(130513)
simulatedData <- simulateData()
standardizedRisks <- with(simulatedData, standardizeRisks(patientCovariates = L, center = center, Y = Y))
head(standardizedRisks)
}
\references{
Varewyck M., Goetghebeur E., Eriksson M. and Vansteelandt S. (2014),
  On shrinkage and model extrapolation in the evaluation of clinical center performance,
  Biostatistics, 15(4), p. 651--664
}

